Enable CUDA profiling compilation, fix warnings

+++ src/acc/cuda/dbcsr_cuda_nvtx_cu.cpp
@@ -12,6 +12,7 @@
 #  include <nvToolsExt.h>
 #  include <stdio.h>
 #  include <pthread.h>
+#  include <cstring>
 
 const uint32_t colormap[] = {0xFFFFFF00, // Yellow
   0xFFFF00FF, // Fuchsia
@@ -39,7 +40,7 @@ extern "C" int cuda_nvtx_range_push_cu(const char* message) {
 
   // colors are picked based on a (very simple) hash value of the message
   int hash = 0;
-  for (int i = 0; i < strlen(message); i++) hash += i * message[i] * message[i];
+  for (size_t i = 0; i < strlen(message); i++) hash += i * message[i] * message[i];
   eventAttrib.colorType = NVTX_COLOR_ARGB;
   eventAttrib.color = colormap[hash % 14];
